requirements.txt
Django==3.0.3
pytest-django==3.8.0
django-debug-toolbar==2.2
python-slugify==4.0.0
Pillow==7.0.0
gunicorn==20.0.4
prettyconf==2.0.1
dj-static==0.0.6
dj-database-url==0.5.0
djangorestframework==3.11.0
psycopg2